% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetaTable.R
\name{createMetaTable}
\alias{createMetaTable}
\title{Convert NEON download to metadata table for Phyloseq package}
\usage{
createMetaTable(neonUtilObject, gene = NA, sampleType = NA)
}
\arguments{
\item{neonUtilObject}{the object downloaded with neonUtilities loadByProduct}

\item{gene}{the marker to process. Either 16S or ITS}

\item{sampleType}{either soil, benthic, or surface}
}
\value{
A table of basic metadata fields for each sample downloaded, extracted from sampleMetadata table. Ready to import to Phyloseq
}
\description{
From Microbe Community Taxonomy tables downloaded from neonUtilities (expanded package), extracts basic fields (e.g. siteID, collectDate) from the SampleMetadata table to a metadata table ready to import to Phyloseq
}
\examples{
\dontrun{
# To convert community taxonomy data object downloaded with neonUtilities to Phyloseq-ready metadata table:
benthic.meta.its <- createMetaTable(benthic.mct, gene='ITS', sampleType = 'benthic')
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Hugh Cross \email{crossh@battelleecology.org}
}
