% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxTable.R
\name{createTaxTable}
\alias{createTaxTable}
\title{Convert NEON download to taxon table for Phyloseq package}
\usage{
createTaxTable(neonUtilObject, gene = NA, sampleType = NA)
}
\arguments{
\item{neonUtilObject}{the object downloaded with neonUtilities loadByProduct}

\item{gene}{the marker to process. Either 16S or ITS}

\item{sampleType}{either soil, benthic, or surface}
}
\value{
a data matrix of all OTUs downloaded and the taxonomic lineage for each. Ready to import to Phyloseq.
}
\description{
From Microbe Community Taxonomy tables downloaded from neonUtilities (expanded package), convert the PerSampleTaxonomy table to a taxon table ready to import to Phyloseq
}
\examples{
\dontrun{
# To convert community taxonomy data object downloaded with neonUtilities to Phyloseq-ready taxon table:
benthic.tax.its <- createTaxTable(benthic.mct, gene='ITS', sampleType = 'benthic')
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Hugh Cross \email{crossh@battelleecology.org}
}
