% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openGOLD.R
\name{openGOLD}
\alias{openGOLD}
\title{Open JGI landing page for NEON metagenomic samples}
\usage{
openGOLD(
  sampleID,
  goldType = "GOLD Sequencing Project ID",
  sourceTable = neon.metaDB
)
}
\arguments{
\item{sampleID}{the dnaSampleID for the sample}

\item{goldType}{defaults}

\item{sourceTable}{defaults}
}
\value{
opens a web page in default browser
}
\description{
This tool will let the user select NEON samples from a provided data table and then automatically open the JGI IMG page for each sample (maximum of ten)
}
\examples{
\dontrun{
# To open a page on jgi GOLD online database :
openGOLD('BONA_013-O-20230710-COMP-DNA1')
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Hugh Cross \email{crossh@battelleecology.org}
}
