% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMapFile.R
\name{parseMapFile}
\alias{parseMapFile}
\title{parseMapFile derive sequencingRunID from taxon table name}
\usage{
parseMapFile(mapFile, gene)
}
\arguments{
\item{mapFile}{the name of the taxon table}

\item{gene}{the gene under study}
}
\value{
the sequencingRunID
}
\description{
Takes a MCT taxon table name and outputs the sequencingRunID from within the name
}
\examples{
\dontrun{
# To get the sequencingRunID from the following taxon table:
taxtable = 'RUSH_19S_36_3318_HKJCFDRX3_ITS_map.csv'
seqrun <- parseMapFile(taxtable,'ITS')
seqrun: 'HKJCFDRX3'
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Hugh Cross \email{crossh@battelleecology.org}
}
