% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitID.R
\name{splitID}
\alias{splitID}
\title{splitID split string into parts by underscore and keep one or series of parts}
\usage{
splitID(string, first, last)
}
\arguments{
\item{string}{the string of characters delimited by underscore}

\item{first}{the first segment to keep after splitting}

\item{last}{the last segment of string to keep after splitting}
}
\value{
a new string with only the first to the last parts collapsed back together with underscore
}
\description{
Takes a string of characters delimited by underscore, breaks it by underscore, and then collapses selected pieces (by number) back
}
\examples{
\dontrun{
# To split up string and retain 5th and 6th parts:
taxtable = 'RUSH_19S_36_3318_HKJCFDRX3_ITS_map.csv'
newtax <- splitID(taxtable,5,6)
newtax: 'HKJCFDRX3_ITS'
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Hugh Cross \email{crossh@battelleecology.org}
}
