% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFreqTable.R
\name{createFreqTable}
\alias{createFreqTable}
\title{Convert NEON download to OTU frequency table for Phyloseq package}
\usage{
createFreqTable(neonUtilObject, gene = NA, sampleType = NA)
}
\arguments{
\item{neonUtilObject}{the object downloaded with neonUtilities loadByProduct}

\item{gene}{the marker to process. Either 16S or ITS}

\item{sampleType}{either soil, benthic, or surface}
}
\value{
A matrix of counts for each sample (columns) for each OTU (rows) from the downloaded data product. Ready to import to Phyloseq
}
\description{
From Microbe Community Taxonomy tables downloaded from neonUtilities (expanded package), convert the PerSampleTaxonomy table to a frequency table ready to import to Phyloseq
}
\examples{
\dontrun{
# To convert community taxonomy data object downloaded with neonUtilities to Phyloseq-ready OTU table:
benthic.freq.its <- createFreqTable(benthic.mct, gene = 'ITS', sampleType = 'benthic')
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Hugh Cross \email{crossh@battelleecology.org}
}
